Use token to create client
- Change methods to use an incoming token
- Create a method to create a client
- Stop using a repository to retrieve tokens

+++ domain/user.go
@@ -2,7 +2,7 @@ package domain
 
 // User is a type where the user attributes are stored
 type User struct {
-	ID          int64
+	ID          int
 	Username    string
 	AccessToken string
 }
